strip vicon and old qualisys client
